fix : modify api
