+++ src/rooter.js
@@ -13,6 +13,8 @@ import {openNavSmartphone} from '../module/navSmartphone.js'
 import {closeNavSmartphone} from '../module/navSmartphone.js'
 // test views system
 import {views} from '../module/views.js' 
+// test api
+import {test} from '../module/pokeApi.js'
 
 // test
 // import {action} from '../module/test.js'
@@ -27,6 +29,7 @@ closeNav();
 openNavSmartphone();
 closeNavSmartphone();
 views();
+test();
 
 // action();
 // action2();
